test: set up multi-level test infrastructure
Add persistence/, engine/, regression/ test packages with fixtures.
Move test_knowledge_system.py to engine/. Add pytest markers.
Add data builder functions for all persistence entities.
