+++ pixi.lock
@@ -2782,7 +2782,7 @@ packages:
 - pypi: ./
   name: session-intelligence
   version: 1.0.0
-  sha256: 54efe367f326f8e6730eb987e5252b597695ae40033f76fa822fba7a09d5ee39
+  sha256: d2a417e6e12cef007f93fff95138b3cc8d489c8440af170b9d06216fa3a87eb9
   requires_dist:
   - pytest>=7.0.0 ; extra == 'dev'
   - pytest-asyncio>=0.21.0 ; extra == 'dev'
