+++ pyproject.toml
@@ -100,6 +100,7 @@ mypy = ">=1.5.0"
 [tool.pixi.feature.quality.pypi-dependencies]
 black = ">=23.0.0"
 build = ">=1.0.0"
+httpx = ">=0.27"
 
 [tool.pixi.feature.quality-extended.dependencies]
 bandit = ">=1.7.0"
@@ -181,6 +182,7 @@ markers = [
     "unit: unit tests for isolated components",
     "asyncio: async tests",
     "postgresql: tests requiring PostgreSQL",
+    "regression: regression tests for previously fixed bugs",
 ]
 filterwarnings = [
     "ignore::DeprecationWarning",
